requirements.txt
torch
torchvision
tqdm
scikit-learn
matplotlib
pyyaml
